backend/blueback/requirements.txt
asn1crypto==0.24.0
astroid==1.6.6
autopep8==1.5
backports.functools-lru-cache==1.6.1
cffi==1.14.0
configparser==4.0.2
cryptography==2.9.2
Django==1.11.29
django-cors-headers==3.0.2
django-filter==2.0.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.9.4
enum34==1.1.10
funcsigs==1.0.2
futures==3.3.0
idna==2.6
ipaddress==1.0.23
isort==4.3.21
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
mock==3.0.5
mysqlclient==1.4.6
pep8==1.7.1
Pillow==6.2.2
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
pylint==1.9.5
pytz==2020.1
pyxdg==0.25
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.2.4
wrapt==1.12.0
